+++ dojo/tools/deepfence_threatmapper/malware.py
@@ -1,0 +1,39 @@
+from dojo.models import Finding
+
+
+class DeepfenceThreatmapperMalware(object):
+    def get_findings(self, row, headers, test):
+        description = ""
+        Rule_Name = row[headers["Rule Name"]]
+        Class = row[headers["Class"]]
+        File_Name = row[headers["File Name"]]
+        Summary = row[headers["Summary"]]
+        Severity = row[headers["Severity"]]
+        Node_Name = row[headers["Node Name"]]
+        NodeType = row[headers["NodeType"]]
+        Container_Name = row[headers["Container Name"]]
+        Kubernetes_Cluster_Name = row[headers["Kubernetes Cluster Name"]]
+        description += "**Summary:** " + str(Summary) + "\n"
+        description += "**Rule Name:** " + str(Rule_Name) + "\n"
+        description += "**Class:** " + str(Class) + "\n"
+        description += "**File Name:** " + str(File_Name) + "\n"
+        description += "**Node Name:** " + str(Node_Name) + "\n"
+        description += "**NodeType:** " + str(NodeType) + "\n"
+        description += "**Container Name:** " + str(Container_Name) + "\n"
+        description += "**Kubernetes Cluster Name:** " + str(Kubernetes_Cluster_Name) + "\n"
+        finding = Finding(
+            title=Rule_Name,
+            description=description,
+            file_path=File_Name,
+            severity=self.severity(Severity),
+            static_finding=False,
+            dynamic_finding=True,
+            test=test,
+        )
+        return finding
+
+    def severity(self, input):
+        if input is None:
+            return "Info"
+        else:
+            return input.capitalize()
